chatbot options only available if auth
